backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
supabase==2.9.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
httpx>=0.27.0,<0.29.0
anthropic==0.40.0
google-auth==2.36.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.154.0
stripe==11.3.0
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.7.0
apscheduler==3.10.4
slowapi==0.1.9
tenacity==9.0.0
sentence-transformers==3.3.1
numpy==1.26.4
